+++ R/jamenrich-import.R
@@ -735,3 +735,4 @@ find_colname <- function
    }
    return(head(x_vals, max));
 }
+